% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colsums_spaces.R
\name{colsums_spaces}
\alias{colsums_spaces}
\title{colsums_spaces}
\usage{
colsums_spaces(table_name)
}
\arguments{
\item{x}{is <numeric> (int or dbl)}
}
\value{
a number that is x + 1 <numeric>
}
\description{
Show no. of cells w/ only spaces in a table column
}
