% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoodBye.R
\name{goodbye}
\alias{goodbye}
\title{goodbye}
\usage{
goodbye(your_name)
}
\arguments{
\item{your_name}{is <string>}
}
\value{
a greeting combining strings
}
\description{
Take input from keyboard and sends "Goodbye <TEXT>"
}
\examples{
> goodbye("Frank")
# [1] "Goodbye Frank"
[1] "Goodbye Frank"
}
